+++ packages/blade/src/components/Carousel/Indicators/StyledIndicatorButton.native.tsx
@@ -4,10 +4,10 @@ import styled from 'styled-components/native';
 import Animated, { useAnimatedStyle, withTiming } from 'react-native-reanimated';
 import { getIndicatorButtonStyles } from './getIndicatorButtonStyles';
 import type { IndicatorButtonProps } from './types';
-import BaseBox from '~components/Box/BaseBox';
 import { castNativeType, makeMotionTime } from '~utils';
-import { useTheme } from '~components/BladeProvider';
 import { size } from '~tokens/global';
+import BaseBox from '~components/Box/BaseBox';
+import { useTheme } from '~components/BladeProvider';
 
 const PressableIndicatorButton = styled(BaseBox)<
   Pick<IndicatorButtonProps, 'variant' | 'isActive'>