+++ packages/blade/src/components/Dropdown/docs/_KitchenSink.Dropdown.stories.tsx
@@ -1,4 +1,4 @@
-import { composeStories } from '@storybook/react';
+import { composeStories } from '@storybook/react-vite';
 import * as dropddownWithButtonStories from './DropdownWithButton.stories';
 import * as dropddownWithSelectStories from './DropdownWithSelect.stories';
 import { Box } from '~components/Box';