+++ packages/blade/src/components/Input/TextInput/_KitchenSink.TextInput.stories.tsx
@@ -1,4 +1,4 @@
-import { composeStories } from '@storybook/react';
+import { composeStories } from '@storybook/react-vite';
 import * as textInputStories from './TextInput.stories';
 import { Box } from '~components/Box';
 import { Heading } from '~components/Typography';