+++ packages/blade/src/components/Menu/docs/_KitchenSink.Menu.stories.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { composeStories } from '@storybook/react';
+import { composeStories } from '@storybook/react-vite';
 import * as menuStories from './Menu.stories';
 import { Box } from '~components/Box';
 import { Heading } from '~components/Typography';