+++ packages/blade/src/components/SideNav/SideNavItems/SideNavItem.web.tsx
@@ -39,12 +39,12 @@ const SideNavItem = ({
         paddingX="spacing.3"
         height={makeSize(NAV_ITEM_HEIGHT)}
         backgroundColor={backgroundColor}
-        borderRadius="medium"
+        borderRadius="small"
         as={as}
         cursor={as === 'label' ? 'pointer' : undefined}
         {...makeAnalyticsAttribute(rest)}
       >
-        <Box display="inline-flex" alignItems="center" gap="spacing.3">
+        <Box display="inline-flex" alignItems="center" gap={Icon ? 'spacing.2' : 'spacing.3'}>
           {Icon ? (
             <BaseBox display="flex" alignItems="center" paddingX="spacing.2">
               <Icon size="medium" color="interactive.icon.gray.subtle" />
@@ -55,7 +55,7 @@ const SideNavItem = ({
           <BaseBox className={classes.HIDE_WHEN_COLLAPSED}>
             <Text
               truncateAfterLines={1}
-              weight="medium"
+              weight="regular"
               size="medium"
               color="surface.text.gray.subtle"
             >