+++ packages/blade-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @razorpay/blade-core
 
+## 0.4.0
+
+### Minor Changes
+
+- 142949f30: feat: blade spark redesign for blade-core and blade-svelte.
+
+## 0.3.0
+
+### Minor Changes
+
+- 46dc37f98: Added Divider component
+
 ## 0.2.0
 
 ### Minor Changes